requirements.txt
pillow==10.0.1
openai==1.3.0
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0